refactor: (#685) common 폴더에 있던 Post를 post 폴더로 이동
