+++ frontend/src/components/common/PostMenu/PostMenu.stories.tsx
@@ -14,6 +14,7 @@ type Story = StoryObj<typeof PostMenu>;
 const menuList: PostMenuItem[] = [
   { color: 'black', content: '닉네임 신고', action: 'NICKNAME_REPORT' },
   { color: 'black', content: '게시글 신고', action: 'POST_REPORT' },
+  { color: 'black', content: '방구', action: 'DELETE' },
 ];
 
 export const Default: Story = {
