+++ frontend/src/components/post/Post/index.tsx
@@ -13,13 +13,12 @@ import { PATH } from '@constants/path';
 import { POST } from '@constants/vote';
 
 import { linkifyText } from '@utils/post/formatTextLink';
-
 import { checkClosedPost, convertTimeToWord } from '@utils/time';
 
 import commentIcon from '@assets/comment.svg';
 import photoIcon from '@assets/photo_black.svg';
 
-import Toast from '../Toast';
+import Toast from '../../common/Toast';
 
 import * as S from './style';
