+++ frontend/src/components/post/Post/mockData.ts
