+++ frontend/src/components/post/Post/style.ts
