+++ frontend/src/components/post/PostList/index.tsx
@@ -7,9 +7,9 @@ import { useIntersectionObserver } from '@hooks/useIntersectionObserver';
 import { usePostRequestInfo } from '@hooks/usePostRequestInfo';
 import { useSelect } from '@hooks/useSelect';
 
-import Post from '@components/common/Post';
 import Select from '@components/common/Select';
 import Skeleton from '@components/common/Skeleton';
+import Post from '@components/post/Post';
 import { SORTING_OPTION, STATUS_OPTION } from '@components/post/PostListPage/constants';
 import type { PostSorting, PostStatus } from '@components/post/PostListPage/types';
